requirements.txt
flask
gunicorn
twilio
python-dotenv
matplotlib
requests
